requirements.txt
pydantic
pydantic-ai
google-genai
elevenlabs
python-dotenv
pdf2image
Pillow
